Tests got crowded, split off base ones
